backend/core-services/requirements.txt
# Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
django-filter>=23.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Authentication
djangorestframework-simplejwt>=5.2.0
python-jose[cryptography]>=3.3.0

# Async/Celery
celery>=5.2.0
redis>=4.3.0

# Storage
django-storages>=1.13.0
boto3>=1.26.0

# Monitoring
sentry-sdk>=1.12.0
django-prometheus>=2.3.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Code Quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
